requirements.txt
azure-servicebus==0.50.0
waitress==1.2.1
flask==1.0.2